refactor: reorganize client structure - move components, hooks, and types to dedicated directories
