+++ client/src/app/about/page.tsx
@@ -1,7 +1,7 @@
 'use client';
 
 import React from 'react';
-import Header from '../header';
+import Header from '../components/layout/header';
 
 function About() {
     return (
