+++ client/src/app/components/layout/header.tsx
