+++ client/src/app/editpost/[id]/page.tsx
@@ -2,7 +2,7 @@
 import { useParams } from "next/navigation";
 import { useEffect, useState } from "react";
 import PostForm from "@/app/newpost/postform";
-import Header from "@/app/header";
+import Header from "@/app/components/layout/header";
 import { fetchPostById } from "../../postview/fetchpostid";
 import { useRouter } from "next/navigation";
 import { getApiUrl } from '../../utils/api';
