+++ client/src/app/page.tsx
@@ -1,4 +1,4 @@
-import Header from "./header";
+import Header from "./components/layout/header";
 
 export default function Home() {
   return (
