+++ client/src/app/postview/[id]/page.tsx
@@ -1,7 +1,7 @@
 "use client"
 import { useParams } from "next/navigation"
-import { useState, useEffect} from "react";
-import Header from "../../header";
+import { useState, useEffect } from "react";
+import Header from "../../components/layout/header";
 import { Post } from "@/types/posts";
 import { Comment } from "@/types/comments";
 import { FormEvent } from "react";
@@ -10,7 +10,7 @@ import EditButton from "@/app/posts/editbutton";
 import CommentForm from "../commentform";
 import UpdateComment from "../updateform";
 import { getApiUrl } from '../../utils/api';
-import useAuth from "../useauth";
+import useAuth from "../../hooks/useauth";
 
 function PostView() {
     const { id } = useParams()
@@ -18,7 +18,7 @@ function PostView() {
     const [post, setPost] = useState<Post | null>(null);
     const [loading, setLoading] = useState(true);
     const [error, setError] = useState<string | null>(null);
-    const {user, token} = useAuth();
+    const { user, token } = useAuth();
     const [comments, setComments] = useState<Comment[]>([]);
     const [commentTitle, setCommentTitle] = useState("");
     const [commentContent, setCommentContent] = useState("");
@@ -28,20 +28,20 @@ function PostView() {
     const [showForm, setShowForm] = useState(false);
     const [activeCommentId, setActiveCommentId] = useState<number | null>(null);
 
-  useEffect(() => {
-  async function loadPost() {
-        try {
-        const data = await fetchPostById(postId);
-        setPost(data);
-        setLoading(false);
-        } catch (error) {
-        setError("Failed to fetch post");
-        setLoading(false);
-        console.error("Error fetching post:", error);
+    useEffect(() => {
+        async function loadPost() {
+            try {
+                const data = await fetchPostById(postId);
+                setPost(data);
+                setLoading(false);
+            } catch (error) {
+                setError("Failed to fetch post");
+                setLoading(false);
+                console.error("Error fetching post:", error);
+            }
         }
-    }
 
-    loadPost();
+        loadPost();
     }, [postId]);
 
 
@@ -62,60 +62,60 @@ function PostView() {
     }, [postId]);
 
     async function postComment(event: FormEvent<HTMLFormElement>) {
-            setError("")
-            setMessage("")
-            setCommentError("")
-            event.preventDefault();
-            const json = {
+        setError("")
+        setMessage("")
+        setCommentError("")
+        event.preventDefault();
+        const json = {
             title: commentTitle,
             content: commentContent,
             postId: Number(postId),
             userId: user?.userid || "",
-            };
-            try {
-                const res = await fetch(getApiUrl(`/comments/post/${postId}`), {
-                    method: "POST",
-                    headers: {
-                        "Content-Type": "application/json",
-                        Authorization: `Bearer ${token}`,
-                    },
-                    body: JSON.stringify(json),
-                });
-                const data = await res.json();
-                if (!res.ok) {
-                    setCommentError(data.message || "Failed to post comment");
-                    return;
-                }
-                setComments((prevComments) => [...prevComments, data]);
-                setMessage("Comment posted successfully!");
-                setCommentLoading(false);
-            } catch (error) {
-                console.error("Error posting comment:", error);
-                setCommentError("Failed to post comment");
+        };
+        try {
+            const res = await fetch(getApiUrl(`/comments/post/${postId}`), {
+                method: "POST",
+                headers: {
+                    "Content-Type": "application/json",
+                    Authorization: `Bearer ${token}`,
+                },
+                body: JSON.stringify(json),
+            });
+            const data = await res.json();
+            if (!res.ok) {
+                setCommentError(data.message || "Failed to post comment");
+                return;
             }
+            setComments((prevComments) => [...prevComments, data]);
+            setMessage("Comment posted successfully!");
+            setCommentLoading(false);
+        } catch (error) {
+            console.error("Error posting comment:", error);
+            setCommentError("Failed to post comment");
         }
+    }
 
-        async function deleteComment(id: number) {
-            try {
-                const res = await fetch(getApiUrl(`/comments/${id}`), {
-                    method: "DELETE",
-                    headers: {
-                        "Content-Type": "application/json",
-                        Authorization: `Bearer ${token}`,
-                    },
-                });
-                const data = await res.json();
-                if (!res.ok) {
-                    setCommentError(data.message || "Failed to delete comment");
-                    return;
-                }
-                setComments((prevComments) => prevComments.filter((comment) => comment.id !== id));
-                setMessage("Comment deleted successfully!");
-            } catch (error) {
-                console.error("Error deleting comment:", error);
-                setCommentError("Failed to delete comment");
+    async function deleteComment(id: number) {
+        try {
+            const res = await fetch(getApiUrl(`/comments/${id}`), {
+                method: "DELETE",
+                headers: {
+                    "Content-Type": "application/json",
+                    Authorization: `Bearer ${token}`,
+                },
+            });
+            const data = await res.json();
+            if (!res.ok) {
+                setCommentError(data.message || "Failed to delete comment");
+                return;
             }
+            setComments((prevComments) => prevComments.filter((comment) => comment.id !== id));
+            setMessage("Comment deleted successfully!");
+        } catch (error) {
+            console.error("Error deleting comment:", error);
+            setCommentError("Failed to delete comment");
         }
+    }
 
 
     if (loading) {
@@ -138,7 +138,7 @@ function PostView() {
                     <p>Updated: {new Date(post.updatedAt).toISOString().split("T")[0]}</p>
                     <p>Status: <span className={post.isPublic ? "text-green-400" : "text-red-400"}>{post.isPublic ? "Public" : "Private"}</span></p>
                 </div>
-                <EditButton id={post.id} isAuthor={user?.isAuthor ?? null}/>
+                <EditButton id={post.id} isAuthor={user?.isAuthor ?? null} />
             </div>
 
             <div className="bg-gray-900 p-6 rounded-lg shadow-md">
@@ -147,26 +147,26 @@ function PostView() {
                 {message && <p className="text-green-400">{message}</p>}
                 {commentError && <p className="text-red-400">{commentError}</p>}
                 {token && (
-                  <>
-                    <button
-                      onClick={() => setShowForm(!showForm)}
-                      className="bg-blue-500 text-white px-4 py-2 rounded-md mb-4"
-                    >
-                      {showForm ? "Hide Form" : "Add Comment"}
-                    </button>
-                    {showForm && (
-                        <CommentForm
-                            postId={Number(postId)}
-                            userId={user?.userid || ""}
-                            onSubmit={postComment}
-                            submitLabel="Add Comment"
-                            title={commentTitle}
-                            content={commentContent}
-                            setTitle={setCommentTitle}
-                            setContent={setCommentContent}
-                        />
-                    )}
-                  </>
+                    <>
+                        <button
+                            onClick={() => setShowForm(!showForm)}
+                            className="bg-blue-500 text-white px-4 py-2 rounded-md mb-4"
+                        >
+                            {showForm ? "Hide Form" : "Add Comment"}
+                        </button>
+                        {showForm && (
+                            <CommentForm
+                                postId={Number(postId)}
+                                userId={user?.userid || ""}
+                                onSubmit={postComment}
+                                submitLabel="Add Comment"
+                                title={commentTitle}
+                                content={commentContent}
+                                setTitle={setCommentTitle}
+                                setContent={setCommentContent}
+                            />
+                        )}
+                    </>
                 )}
 
                 {!token && (
@@ -190,38 +190,38 @@ function PostView() {
                             <p><span className="text-sm text-gray-400">Created:</span> {new Date(comment.createdAt).toISOString().split("T")[0]}</p>
                             <p><span className="text-sm text-gray-400">Updated:</span> {new Date(comment.updatedAt).toISOString().split("T")[0]}</p>
                             {comment.user?.username && (
-                              <p><span className="text-sm text-gray-400">Author:</span> {comment.user.username}</p>
+                                <p><span className="text-sm text-gray-400">Author:</span> {comment.user.username}</p>
                             )}
                             <p><span className="font-semibold">Role:</span> {comment.user?.isAuthor ? 'Author' : 'User'}</p>
                             {comment.user && comment.user.userid === user?.userid && (
-                              <div className="flex flex-col gap-2 mt-2">
-                                <button
-                                  className="text-blue-400 hover:underline"
-                                  onClick={() => 
-                                    setActiveCommentId(activeCommentId === comment.id ? null : comment.id)}
-                                >
-                                  {activeCommentId === comment.id ? "Hide Update Form" : "Edit Comment"}
-                                </button>
-                                {activeCommentId === comment.id && (
-                                  <UpdateComment
-                                    comment={comment}
-                                    token={token ? token : ""}
-                                  />
-                                )}
-                                <button
-                                  className="text-red-400 hover:underline"
-                                  onClick={() => deleteComment(comment.id)}
-                                >
-                                  Delete Comment
-                                </button>
-                              </div>
+                                <div className="flex flex-col gap-2 mt-2">
+                                    <button
+                                        className="text-blue-400 hover:underline"
+                                        onClick={() =>
+                                            setActiveCommentId(activeCommentId === comment.id ? null : comment.id)}
+                                    >
+                                        {activeCommentId === comment.id ? "Hide Update Form" : "Edit Comment"}
+                                    </button>
+                                    {activeCommentId === comment.id && (
+                                        <UpdateComment
+                                            comment={comment}
+                                            token={token ? token : ""}
+                                        />
+                                    )}
+                                    <button
+                                        className="text-red-400 hover:underline"
+                                        onClick={() => deleteComment(comment.id)}
+                                    >
+                                        Delete Comment
+                                    </button>
+                                </div>
                             )}
-                        </div>  
+                        </div>
                     ))}
                 </div>
             </div>
         </>
     )
-    
+
 }
 export default PostView;
