--- client/src/app/types.ts
@@ -1,10 +1,0 @@
-export interface User {
-    name: string;
-}
-
-export interface Post {
-    id: number;
-    title: string;
-    content: string;
-    user: User;
-}
